asdasd

+++ Pr_3/src/c/multiplicarMatrices.c
@@ -3,7 +3,7 @@
 #include <stdlib.h>
 #include <sys/time.h>
 
-#include "../h/arqo3.h"
+#include "arqo3.h"
 
 void compute(tipo **m1,tipo **m2, tipo **res, int n);
 
@@ -52,7 +52,7 @@ void compute(tipo **m1,tipo **m2, tipo **res, int n) {
 		for(j=0;j<n;j++) {
       resultado = (tipo)0;
       for(k=0;k<n;k++) {
-        resultado += (int)m1[i][k]*(int)m2[k][j];
+        resultado += m1[i][k]*m2[k][j];
       }
 			res[i][j]=resultado;
 		}
